mediscan-backend/mediscan_api/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.7
Pillow==10.0.0
djangorestframework==3.14.0
PyYAML==6.0.1
requests==2.31.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0